+++ inc/stm32f411e_discovery.h
@@ -1,39 +1,3 @@
-/** 
-  ******************************************************************************
-  * @file    stm32f411e_discovery.h
-  * @author  MCD Application Team
-  * @brief   This file contains definitions for STM32F401-Discovery Kit's Leds and 
-  *          push-button hardware resources.
-  ******************************************************************************
-  * @attention
-  *
-  * <h2><center>&copy; COPYRIGHT(c) 2017 STMicroelectronics</center></h2>
-  *
-  * Redistribution and use in source and binary forms, with or without modification,
-  * are permitted provided that the following conditions are met:
-  *   1. Redistributions of source code must retain the above copyright notice,
-  *      this list of conditions and the following disclaimer.
-  *   2. Redistributions in binary form must reproduce the above copyright notice,
-  *      this list of conditions and the following disclaimer in the documentation
-  *      and/or other materials provided with the distribution.
-  *   3. Neither the name of STMicroelectronics nor the names of its contributors
-  *      may be used to endorse or promote products derived from this software
-  *      without specific prior written permission.
-  *
-  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-  * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
-  * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
-  * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
-  * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
-  * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
-  * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-  *
-  ******************************************************************************  
-  */ 
-  
 /* Define to prevent recursive inclusion -------------------------------------*/
 #ifndef __STM32F411E_DISCOVERY_H
 #define __STM32F411E_DISCOVERY_H
@@ -44,77 +8,28 @@
                                               
 /* Includes ------------------------------------------------------------------*/
 #include "stm32f4xx_hal.h"
-   
-/** @addtogroup BSP
-  * @{
-  */
-  
-/** @addtogroup STM32F411E_DISCOVERY
-  * @{
-  */
-      
-/** @defgroup __STM32F411E_DISCOVERY_HLOW_LEVEL STM32F411E DISCOVERY HLOW LEVEL
-  * @{
-  */ 
+#include "bsp.h"
 
-/** @defgroup __STM32F411E_DISCOVERY_HLOW_LEVEL_Exported_Types STM32F411E DISCOVERY HLOW LEVEL Exported Types
-  * @{
-  */
-typedef enum 
-{
-  LED4 = 0,
-  LED3 = 1,
-  LED5 = 2,
-  LED6 = 3
-}Led_TypeDef;
-
-typedef enum 
-{  
-  BUTTON_KEY = 0,
-}Button_TypeDef;
-
-typedef enum 
-{  
-  BUTTON_MODE_GPIO = 0,
-  BUTTON_MODE_EXTI = 1
-}ButtonMode_TypeDef;     
-/**
-  * @}
-  */ 
-
-/** @defgroup __STM32F411E_DISCOVERY_HLOW_LEVEL_Exported_Constants STM32F411E DISCOVERY HLOW LEVEL Exported Constants
-  * @{
-  */ 
-
-/** 
-  * @brief Define for STM32F411E_DISCOVERY board  
-  */ 
 #if !defined (USE_STM32F411E_DISCO)
  #define USE_STM32F411E_DISCO
 #endif
 
-/** @defgroup __STM32F411E_DISCOVERY_HLOW_LEVEL_LED STM32F411E DISCOVERY HLOW LEVEL LED
-  * @{
-  */
 #define LEDn                                    4
 
 #define LED4_PIN                                GPIO_PIN_12
 #define LED4_GPIO_PORT                          GPIOD
 #define LED4_GPIO_CLK_ENABLE()                  __HAL_RCC_GPIOD_CLK_ENABLE()  
 #define LED4_GPIO_CLK_DISABLE()                 __HAL_RCC_GPIOD_CLK_DISABLE()  
-
   
 #define LED3_PIN                                GPIO_PIN_13
 #define LED3_GPIO_PORT                          GPIOD
 #define LED3_GPIO_CLK_ENABLE()                  __HAL_RCC_GPIOD_CLK_ENABLE()  
 #define LED3_GPIO_CLK_DISABLE()                 __HAL_RCC_GPIOD_CLK_DISABLE()  
 
-  
 #define LED5_PIN                                GPIO_PIN_14
 #define LED5_GPIO_PORT                          GPIOD
 #define LED5_GPIO_CLK_ENABLE()                  __HAL_RCC_GPIOD_CLK_ENABLE()  
 #define LED5_GPIO_CLK_DISABLE()                 __HAL_RCC_GPIOD_CLK_DISABLE()  
-
   
 #define LED6_PIN                                GPIO_PIN_15
 #define LED6_GPIO_PORT                          GPIOD
@@ -132,13 +47,8 @@ typedef enum
                                             if((__INDEX__) == 2) LED5_GPIO_CLK_DISABLE(); else \
                                             if((__INDEX__) == 3) LED6_GPIO_CLK_DISABLE(); \
                                             }while(0)
-/**
-  * @}
-  */ 
-  
-/** @defgroup __STM32F411E_DISCOVERY_HLOW_LEVEL_BUTTON STM32F411E DISCOVERY HLOW LEVEL BUTTON
-  * @{
-  */  
+
+
 #define BUTTONn                                 1  
 
 /**
@@ -154,13 +64,6 @@ typedef enum
                                                  }while(0)
 #define BUTTONx_GPIO_CLK_DISABLE(__INDEX__)    do{if((__INDEX__) == 0) KEY_BUTTON_GPIO_CLK_DISABLE(); \
                                                  }while(0)
-/**
-  * @}
-  */ 
-
-/** @defgroup __STM32F411E_DISCOVERY_HLOW_LEVEL_BUS  STM32F411E DISCOVERY HLOW LEVEL BUS
-  * @{
-  */  
 
 /*############################### I2Cx #######################################*/
 #define DISCOVERY_I2Cx                          I2C1
@@ -264,48 +167,11 @@ typedef enum
 #define ACCELERO_INT1_EXTI_IRQn                 EXTI4_IRQn 
 #define ACCELERO_INT2_PIN                       GPIO_PIN_5                  /* PE.05 */
 #define ACCELERO_INT2_EXTI_IRQn                 EXTI9_5_IRQn 
-/**
-  * @}
-  */ 
 
-/** @defgroup __STM32F411E_DISCOVERY_HLOW_LEVEL_Exported_Macros STM32F411E DISCOVERY HLOW LEVEL Exported Macros
-  * @{
-  */  
-/**
-  * @}
-  */ 
 
-/** @defgroup __STM32F411E_DISCOVERY_HLOW_LEVEL_Exported_Functions  STM32F411E DISCOVERY HLOW LEVEL Exported Functions
-  * @{
-  */
-uint32_t BSP_GetVersion(void);  
-void     BSP_LED_Init(Led_TypeDef Led);
-void     BSP_LED_On(Led_TypeDef Led);
-void     BSP_LED_Off(Led_TypeDef Led);
-void     BSP_LED_Toggle(Led_TypeDef Led);
-void     BSP_PB_Init(Button_TypeDef Button, ButtonMode_TypeDef ButtonMode);
-uint32_t BSP_PB_GetState(Button_TypeDef Button);
-
-/**
-  * @}
-  */
-
-/**
-  * @}
-  */ 
-
-/**
-  * @}
-  */
-
-/**
-  * @}
-  */ 
 
 #ifdef __cplusplus
 }
 #endif
 
 #endif /* __STM32F411E_DISCOVERY_H */
-
-/************************ (C) COPYRIGHT STMicroelectronics *****END OF FILE****/
